refactor: Extract DatePickerField
