+++ app/configurator/components/field-date-picker.tsx
@@ -1,0 +1,168 @@
+import { DatePicker, DatePickerProps } from "@mui/lab";
+import { DatePickerView } from "@mui/lab/DatePicker/shared";
+import { Box, TextField } from "@mui/material";
+import { timeFormat } from "d3-time-format";
+import React from "react";
+
+import { Label } from "@/components/form";
+import { TimeUnit } from "@/graphql/resolver-types";
+
+const formatDate = timeFormat("%Y-%m-%d");
+
+type DatePickerTimeUnit =
+  | TimeUnit.Day
+  | TimeUnit.Week
+  | TimeUnit.Month
+  | TimeUnit.Year;
+
+export const canRenderDatePickerField = (
+  timeUnit: TimeUnit
+): timeUnit is DatePickerTimeUnit => {
+  return [TimeUnit.Day, TimeUnit.Week, TimeUnit.Month, TimeUnit.Year].includes(
+    timeUnit
+  );
+};
+
+const getDateRenderProps = (
+  timeUnit: DatePickerTimeUnit,
+  isDateDisabled: (d: any) => boolean
+):
+  | { shouldDisableDate: DatePickerProps["shouldDisableDate"] }
+  | { shouldDisableYear: DatePickerProps["shouldDisableYear"] } => {
+  switch (timeUnit) {
+    case TimeUnit.Day:
+    case TimeUnit.Week:
+    case TimeUnit.Month:
+      return { shouldDisableDate: isDateDisabled };
+    case TimeUnit.Year:
+      return { shouldDisableYear: isDateDisabled };
+    default:
+      const _exhaustiveCheck: never = timeUnit;
+      return _exhaustiveCheck;
+  }
+};
+
+const getInputFormat = (timeUnit: DatePickerTimeUnit): string => {
+  switch (timeUnit) {
+    case TimeUnit.Day:
+    case TimeUnit.Week:
+      return "dd.MM.yyyy";
+    case TimeUnit.Month:
+      return "MM.yyyy";
+    case TimeUnit.Year:
+      return "yyyy";
+    default:
+      const _exhaustiveCheck: never = timeUnit;
+      return _exhaustiveCheck;
+  }
+};
+
+const getViews = (timeUnit: DatePickerTimeUnit): DatePickerView[] => {
+  switch (timeUnit) {
+    case TimeUnit.Day:
+    case TimeUnit.Week:
+      return ["year", "month", "day"];
+    case TimeUnit.Month:
+      return ["year", "month"];
+    case TimeUnit.Year:
+      return ["year"];
+    default:
+      const _exhaustiveCheck: never = timeUnit;
+      return _exhaustiveCheck;
+  }
+};
+
+export const DatePickerField = ({
+  label,
+  name,
+  value,
+  isDateDisabled,
+  onChange,
+  disabled,
+  controls,
+  timeUnit = TimeUnit.Day,
+  dateFormat = formatDate,
+  ...props
+}: {
+  name: string;
+  value: Date;
+  isDateDisabled: (date: Date) => boolean;
+  onChange: (event: React.ChangeEvent<HTMLSelectElement>) => void;
+  controls?: React.ReactNode;
+  label?: string | React.ReactNode;
+  disabled?: boolean;
+  timeUnit?: DatePickerTimeUnit;
+  dateFormat?: (d: Date) => string;
+} & Omit<
+  DatePickerProps<Date>,
+  | "onChange"
+  | "value"
+  | "shouldDisableDate"
+  | "onChange"
+  | "inputFormat"
+  | "renderInput"
+>) => {
+  const handleChange = React.useCallback(
+    (date: Date | null) => {
+      if (!date) {
+        return;
+      }
+
+      const isDisabled = isDateDisabled(date);
+      if (isDisabled) {
+        return;
+      }
+
+      const ev = {
+        target: {
+          value: dateFormat(date),
+        },
+      } as React.ChangeEvent<HTMLSelectElement>;
+
+      onChange(ev);
+    },
+    [isDateDisabled, onChange, dateFormat]
+  );
+
+  const dateLimitProps = getDateRenderProps(timeUnit, isDateDisabled);
+
+  return (
+    <Box
+      sx={{
+        color: disabled ? "grey.300" : "grey.700",
+        fontSize: "1rem",
+      }}
+    >
+      {label && name && (
+        <Label htmlFor={name} smaller sx={{ my: 1 }}>
+          {label}
+          {controls}
+        </Label>
+      )}
+      <DatePicker<Date>
+        {...props}
+        {...dateLimitProps}
+        inputFormat={getInputFormat(timeUnit)}
+        views={getViews(timeUnit)}
+        value={value}
+        onChange={handleChange}
+        onYearChange={handleChange}
+        renderInput={(params) => (
+          <TextField
+            hiddenLabel
+            size="small"
+            {...params}
+            sx={{
+              ...params.sx,
+
+              "& input": {
+                typography: "body2",
+              },
+            }}
+          />
+        )}
+        disabled={disabled}
+      />
+    </Box>
+  );
+};
